requirements.txt
PySide6>=6.6.0
Pillow>=10.0.0
pytest>=8.0.0
pytest-qt>=4.4.0
pyinstaller>=6.0.0
